gecko/dom/usb/test/test_usb_basics.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for USB API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for USB API **/

ok('usb' in navigator, "navigator.usb should exist");

/*
 * TODO: add test cases for deviceAttached/deviceConfgiured attributes.
 */
</script>
</pre>
</body>
</html>
